Implement React-Select Dropdown and refactor component
In order to solve 558, which was an issue with iOS devices (typically
phones) needing double taps to open dropdowns when another one is open,
React-Select was implemented. Sporting a customizable interface,
TypeScript integration, and total ease of use with iPhones to avoid the
aforementioned issue, it seemed like an excellent pivot.

This adjustment is important because the need to double-click one
dropdown for it to open when another dropdown was already open, as in
the TripOptions component, posed accessibility issues and a UX pain
point for users with iPhones

At the moment, styling and UI consulting is needed still.

OTP-UI 558 #time 4d