fix(core-utils): optimize flex combination generation
This accomplishes two things. First, it only allows FLEX_DIRECT on its own, without transit or anything else. Although this should work, OTP has a bug that causes issues with this configuration. It doesn't make sense to combine them anyway.

Second, it makes FLEX (access/egress) *only* comaptible with TRANSIT. As far as I can tell, OTP doesn't support FLEX combined with bikeshare or anything else. This should reduce the number of queries we send to OTP, which will improve performance.

Lastly, it improves the test output to make it easier to read.