fix(itinerary-body): Eliminate erroneous warn logs
The `getPlaceName` util in the itinerary-body package makes a call to
`getCompanyForNetwork` to get a validated company name for micromobility
legs. The latter function logs a warning when the supplied network is
not in the otp-ui company config. Howwever `getPlaceName` is also used
for walk legs, which never have an associated network so this warning is
emitted for each walk leg, each time the itinerary is rendered.  This
commit alters the call to `getCompanyFromNetwork` to only be made when
the place as a truthy network value and eliminates the invalid warnings